Extract action functions from ExAdmin to ExAdmin.ResourceTitleActions
