+++ lib/ex_admin/register.ex
@@ -186,7 +186,7 @@ defmodule ExAdmin.Register do
 
       defstruct controller: @controller,
                 controller_methods: Module.get_attribute(__MODULE__, :controller_methods),
-                title_actions: &ExAdmin.default_resource_title_actions/2,
+                title_actions: &ExAdmin.ResourceTitleActions.default/2,
                 type: :resource,
                 resource_model: module,
                 resource_name: resource_name(module),
