+++ test/resource_title_actions_test.exs
@@ -15,7 +15,7 @@ defmodule TestExAdmin.ExAdmin.SimpleCustom do
   end
 end
 
-defmodule ExAdminTest do
+defmodule ExAdmin.ResourceTitleActionsTest do
   use ExUnit.Case, async: true
 
   setup config do
@@ -33,42 +33,42 @@ defmodule ExAdminTest do
 
   @tag as_resource: %TestExAdmin.ExAdmin.Simple{}
   test "action_button", %{defn: defn, conn: conn} do
-    result = ExAdmin.action_button(conn, defn, "Simple", :show, :edit, defn.actions, "17")
+    result = ExAdmin.ResourceTitleActions.action_button(conn, defn, "Simple", :show, :edit, defn.actions, "17")
     assert result == [{"Edit Simple", [href: "/admin/simples/1/edit"]}]
 
-    result = ExAdmin.action_button(conn, defn, "Simple", :show, :new, defn.actions, "17")
+    result = ExAdmin.ResourceTitleActions.action_button(conn, defn, "Simple", :show, :new, defn.actions, "17")
     assert result == [{"New Simple", [href: "/admin/simples/new"]}]
 
-    result = ExAdmin.action_button(conn, defn, "Simple", :show, :delete, defn.actions, "17")
+    result = ExAdmin.ResourceTitleActions.action_button(conn, defn, "Simple", :show, :delete, defn.actions, "17")
     assert result == [{"Delete Simple", [href: "/admin/simples/1", "data-confirm": "Are you sure you want to delete this?", "data-method": :delete, rel: :nofollow]}]
 
-    result = ExAdmin.action_button(conn, defn, "Simple", :index, :new, defn.actions, "17")
+    result = ExAdmin.ResourceTitleActions.action_button(conn, defn, "Simple", :index, :new, defn.actions, "17")
     assert result == [{"New Simple", [href: "/admin/simples/new"]}]
 
-    result = ExAdmin.action_button(conn, defn, "Simple", :edit, :new, defn.actions, "17")
+    result = ExAdmin.ResourceTitleActions.action_button(conn, defn, "Simple", :edit, :new, defn.actions, "17")
     assert result == [{"New Simple", [href: "/admin/simples/new"]}]
   end
 
   @tag as_resource: %TestExAdmin.ExAdmin.Simple{}
-  test "default_resource_title_actions", %{defn: defn, conn: conn} do
+  test "default actions", %{defn: defn, conn: conn} do
     conn = struct(conn, private: %{phoenix_action: :show})
-    result = ExAdmin.default_resource_title_actions(conn, defn)
+    result = ExAdmin.ResourceTitleActions.default(conn, defn)
     assert result ==  [edit: [{"Edit Simple", [href: "/admin/simples/1/edit"]}], new: [{"New Simple", [href: "/admin/simples/new"]}], delete: [{"Delete Simple", [href: "/admin/simples/1", "data-confirm": "Are you sure you want to delete this?", "data-method": :delete, rel: :nofollow]}]]
 
     conn = struct(conn, private: %{phoenix_action: :index})
-    result = ExAdmin.default_resource_title_actions(conn, defn)
+    result = ExAdmin.ResourceTitleActions.default(conn, defn)
 
     assert result == [new: [{"New Simple", [href: "/admin/simples/new"]}]]
     conn = struct(conn, private: %{phoenix_action: :edit})
 
-    result = ExAdmin.default_resource_title_actions(conn, defn)
+    result = ExAdmin.ResourceTitleActions.default(conn, defn)
     assert result == [new: [{"New Simple", [href: "/admin/simples/new"]}]]
   end
 
   @tag as_resource: %TestExAdmin.ExAdmin.SimpleCustom{}
-  test "default_resource_title_actions custom actions", %{defn: defn, conn: conn} do
+  test "default custom actions", %{defn: defn, conn: conn} do
     conn = struct(conn, private: %{phoenix_action: :index})
-    result = ExAdmin.default_resource_title_actions(conn, defn)
+    result = ExAdmin.ResourceTitleActions.default(conn, defn)
     assert result ==  [
       new: [{"New Simple", [href: "/admin/simples/new"]}],
       custom: [{"Public Bulk", [href: "/admin/simples/collection/public_bulk"]}],
@@ -76,7 +76,7 @@ defmodule ExAdminTest do
     ]
 
     conn = struct(conn, private: %{phoenix_action: :show})
-    result = ExAdmin.default_resource_title_actions(conn, defn)
+    result = ExAdmin.ResourceTitleActions.default(conn, defn)
     assert result ==  [
       edit: [{"Edit Simple", [href: "/admin/simples/1/edit"]}],
       new: [{"New Simple", [href: "/admin/simples/new"]}],
